requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
boto3==1.20.4
botocore==1.23.4
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.9
django-celery-beat==2.3.0
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-heroku==0.3.1
django-redis==5.0.0
django-redis-cache==2.1.1
django-resized==0.3.11
django-storages==1.12.3
django-timezone-field==5.0
Django-Verify-Email==2.0.3
docopt==0.4.0
filelock==3.4.0
h11==0.12.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
kombu==5.2.4
mailchimp==2.0.10
mailchimp-marketing==3.0.75
outcome==1.1.0
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyperclip==1.8.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stripe==2.61.0
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2022.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
whitenoise==5.3.0
wsproto==1.0.0
xerox==0.4.1
zope.interface==5.4.0